+++ apps/web/client/src/components/ui/settings-modal/domain/custom/record-field.tsx
@@ -1,9 +1,40 @@
+import type { VerificationRecord } from '@onlook/models';
 import { Button } from '@onlook/ui/button';
 import { Icons } from '@onlook/ui/icons';
 import { cn } from '@onlook/ui/utils';
-import { useState } from 'react';
+import { observer } from 'mobx-react-lite';
+import { Fragment, useState } from 'react';
+import { useDomainVerification } from './use-domain-verification';
 
-export function RecordField({
+export const DnsRecords = observer(() => {
+    const { verification } = useDomainVerification();
+
+    const txtRecord = verification?.txtRecord;
+    const aRecords = verification?.aRecords ?? [];
+    const records: VerificationRecord[] = [txtRecord, ...aRecords].filter((record) => !!record);
+
+    if (records.length === 0) {
+        return null;
+    }
+
+    return (
+        <div className="grid grid-cols-7 gap-4 rounded-lg border p-4">
+            <div className="text-sm font-medium col-span-1">Type</div>
+            <div className="text-sm font-medium col-span-3">Host</div>
+            <div className="text-sm font-medium col-span-3">Value</div>
+
+            {records.map((record, index) => (
+                <Fragment key={`${record.type}-${record.name}-${index}`}>
+                    <RecordField value={record.type} className="col-span-1" copyable={false} />
+                    <RecordField value={record.name} className="col-span-3" />
+                    <RecordField value={record.value} className="col-span-3" />
+                </Fragment>
+            ))}
+        </div>
+    );
+});
+
+function RecordField({
     value,
     className,
     copyable = true,
